SDR-241 refactor: UserLocationInfo 도메인 객체를 추가하여 검증의 책임을 이동
